Refactoring ahead of string declarations.
In order for a single add by string function to define cells,
docks and wraps to function, all portions of the layout must
be defined as part of the Layout object.

In order to avoid confusing the user with optional parameters,
separate Cell and Dock structs are used for Add and Dock
respectively.

These are converted to the combined layout struct which now
contains the ID and the embedded Cell or Dock object. The
Layout object is now private to avoid adding complexity to
the API.